requirements.txt
python-dotenv
pandas 
numpy
pymongo
certifi
pymongo[srv]==4.6.0
scikit-learn
dill
pyaml
mlflow
##-e .